Improve Rx pipelines
